feat(court): update court managment for owner

+++ scb-backend/src/main/java/com/example/scbbackend/modules/court/dto/response/PriceTemplateSummaryResponse.java
@@ -4,6 +4,6 @@ public record PriceTemplateSummaryResponse(
         Long id,
         String name,
         int version,
-        int appliedCourtCount,
+        long appliedCourtCount,
         boolean isActive
 ) {}
